requirements.txt
numpy==2.1.1
opencv-python==4.10.0.84
pdf2image==1.17.0
pillow==10.4.0
pytesseract==0.3.13
